Move CategoryController to Http directory and add Swagger documentation
- Transferred CategoryController to the Http directory for improved structure.
- Included Swagger documentation for the categories endpoint to enhance API clarity.
